\name{judges_types}
\alias{judges_types}
\docType{data}
\title{
Classification of judges by type
}
\description{
This set contains information about the classification of judges by `judge type'. Relied on by \code{add_background_type_to_appointments} and not intended for direct use.
}
\usage{
}
\details{
For more information about the judge type variable, see our publication "The Ideal Justice" in TfR).
}
\source{
}
\author{
Johan Lindholm, Umeå University, \email{johan.lindholm@umu.se}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{

}

\seealso{
}

\keyword{datasets}
